feat: enhance user experience with motivational quotes, refactor user data handling, and improve workout difficulty mapping
