+++ src/data/mockData.ts
@@ -1,255 +1,34 @@
-import type { FrontendAvatar, FrontendBadge, FrontendExercise, FrontendFriend } from "../types";
-
-export const avatars: FrontendAvatar[] = [
-  {
-    id: "1",
-    name: "Sporty Spice",
-    image: "https://api.dicebear.com/6.x/avataaars/svg?seed=Sporty",
-    price: 100,
-    category: "sports",
-    rarity: "common",
-  },
-  {
-    id: "2",
-    name: "Zen Master",
-    image: "https://api.dicebear.com/6.x/avataaars/svg?seed=Zen",
-    price: 150,
-    category: "lifestyle",
-    rarity: "common",
-  },
-  {
-    id: "3",
-    name: "Power Lifter",
-    image: "https://api.dicebear.com/6.x/avataaars/svg?seed=Power",
-    price: 200,
-    category: "sports",
-    rarity: "rare",
-  },
-  {
-    id: "4",
-    name: "Yoga Guru",
-    image: "https://api.dicebear.com/6.x/avataaars/svg?seed=Yoga",
-    price: 250,
-    category: "lifestyle",
-    rarity: "rare",
-  },
-  {
-    id: "5",
-    name: "Runner",
-    image: "https://api.dicebear.com/6.x/avataaars/svg?seed=Runner",
-    price: 300,
-    category: "sports",
-    rarity: "rare",
-  },
-  {
-    id: "6",
-    name: "Swimmer",
-    image: "https://api.dicebear.com/6.x/avataaars/svg?seed=Swimmer",
-    price: 350,
-    category: "sports",
-    rarity: "epic",
-  },
-  {
-    id: "7",
-    name: "Cyclist",
-    image: "https://api.dicebear.com/6.x/avataaars/svg?seed=Cyclist",
-    price: 400,
-    category: "sports",
-    rarity: "epic",
-  },
-  {
-    id: "8",
-    name: "Boxer",
-    image: "https://api.dicebear.com/6.x/avataaars/svg?seed=Boxer",
-    price: 450,
-    category: "sports",
-    rarity: "legendary",
-  },
-];
+import type { FrontendBadge } from "../types";
 
 export const badges: FrontendBadge[] = [
   {
     id: "1",
     name: "Early Bird",
     icon: "🌅",
-    description: "Completed 5 workouts before 8 AM",
+    description: "Completed 5 workouts before 8 AM"
   },
   {
     id: "2",
     name: "Night Owl",
     icon: "🦉",
-    description: "Completed 10 workouts after 8 PM",
+    description: "Completed 10 workouts after 8 PM"
   },
   {
     id: "3",
     name: "Consistency King",
     icon: "👑",
-    description: "Maintained a 7-day workout streak",
+    description: "Maintained a 7-day workout streak"
   },
   {
     id: "4",
     name: "Muscle Master",
     icon: "💪",
-    description: "Completed 20 strength training workouts",
+    description: "Completed 20 strength training workouts"
   },
   {
     id: "5",
     name: "Cardio Crusher",
     icon: "🏃",
-    description: "Burned over 5000 calories in cardio exercises",
-  },
-];
-
-export const friends: FrontendFriend[] = [
-  {
-    id: 1,
-    name: "Alice Johnson",
-    avatar: "https://api.dicebear.com/6.x/avataaars/svg?seed=Alice",
-    email: "alice@example.com",
-    goals: ["Gain Muscle", "Improve Flexibility"],
-    level: 5,
-  },
-  {
-    id: 2,
-    name: "Bob Smith",
-    avatar: "https://api.dicebear.com/6.x/avataaars/svg?seed=Bob",
-    email: "bob@example.com",
-    goals: ["Lose Weight", "Increase Stamina"],
-    level: 3,
-  },
-  {
-    id: 3,
-    name: "Charlie Brown",
-    avatar: "https://api.dicebear.com/6.x/avataaars/svg?seed=Charlie",
-    email: "charlie@example.com",
-    goals: ["Improve Flexibility", "Increase Stamina"],
-    level: 4,
-  },
-  {
-    id: 4,
-    name: "Diana Prince",
-    avatar: "https://api.dicebear.com/6.x/avataaars/svg?seed=Diana",
-    email: "diana@example.com",
-    goals: ["Gain Muscle", "Lose Weight"],
-    level: 6,
-  },
-];
-
-export const initialExercises: FrontendExercise[] = [
-  {
-    id: "pushups",
-    name: "Push-ups",
-    targetSets: 3,
-    targetReps: 10,
-    restTime: 60,
-    instruction: "Keep your body straight, lower until your chest nearly touches the floor, then push back up.",
-    difficulty: "intermediate",
-    muscleGroup: "Chest, Shoulders, Triceps",
-    xpReward: 100,
-    sets: Array(3).fill({ completed: false, reps: 0 }),
-    currentSet: 0,
-    currentReps: 0,
-    tips: [
-      "Keep your core tight throughout the movement",
-      "Don't let your hips sag or pike up",
-      "Breathe out as you push up",
-      "For easier version, do push-ups on your knees",
-    ],
-  },
-  {
-    id: "squats",
-    name: "Squats",
-    targetSets: 3,
-    targetReps: 15,
-    restTime: 45,
-    instruction:
-      "Stand with feet shoulder-width apart, lower your body as if sitting in a chair, then return to standing.",
-    difficulty: "beginner",
-    muscleGroup: "Quadriceps, Hamstrings, Glutes",
-    xpReward: 120,
-    sets: Array(3).fill({ completed: false, reps: 0 }),
-    currentSet: 0,
-    currentReps: 0,
-    tips: [
-      "Keep your chest up and back straight",
-      "Push your knees outward as you descend",
-      "Go as low as comfortable, ideally thighs parallel to ground",
-      "Push through your heels when standing up",
-    ],
-  },
-  {
-    id: "plank",
-    name: "Plank",
-    targetSets: 3,
-    targetReps: 30, // seconds
-    restTime: 30,
-    instruction: "Hold a push-up position with your weight on your forearms, keeping your body in a straight line.",
-    difficulty: "intermediate",
-    muscleGroup: "Core, Shoulders",
-    xpReward: 150,
-    sets: Array(3).fill({ completed: false, reps: 0 }),
-    currentSet: 0,
-    currentReps: 0,
-    tips: [
-      "Keep your shoulders directly above your elbows",
-      "Engage your core and glutes",
-      "Don't let your hips sag or pike up",
-      "Look slightly forward, not straight down",
-    ],
-  },
-  {
-    id: "lunges",
-    name: "Lunges",
-    targetSets: 3,
-    targetReps: 12,
-    restTime: 45,
-    instruction:
-      "Step forward with one leg, lowering your hips until both knees are bent at 90 degrees, then return to standing.",
-    difficulty: "intermediate",
-    muscleGroup: "Quadriceps, Hamstrings, Glutes",
-    xpReward: 120,
-    sets: Array(3).fill({ completed: false, reps: 0 }),
-    currentSet: 0,
-    currentReps: 0,
-    tips: [
-      "Keep your upper body straight",
-      "Step far enough forward that your knee stays above your ankle",
-      "Push back up through your front heel",
-      "Alternate legs for each rep",
-    ],
-  },
-  {
-    id: "mountainClimbers",
-    name: "Mountain Climbers",
-    targetSets: 3,
-    targetReps: 20,
-    restTime: 30,
-    instruction: "Start in a plank position and alternate bringing each knee toward your chest in a running motion.",
-    difficulty: "intermediate",
-    muscleGroup: "Core, Shoulders, Hip Flexors",
-    xpReward: 130,
-    sets: Array(3).fill({ completed: false, reps: 0 }),
-    currentSet: 0,
-    currentReps: 0,
-    tips: [
-      "Keep your hips down and core engaged",
-      "Move your legs as quickly as you can while maintaining form",
-      "Breathe rhythmically throughout the exercise",
-      "Each knee drive counts as one rep",
-    ],
-  },
-];
-
-// Motivational quotes for fitness
-export const motivationalQuotes = [
-  "The only bad workout is the one that didn't happen.",
-  "Your body can stand almost anything. It's your mind that you have to convince.",
-  "The pain you feel today will be the strength you feel tomorrow.",
-  "Fitness is not about being better than someone else. It's about being better than you used to be.",
-  "The hardest lift of all is lifting your butt off the couch.",
-  "Don't wish for it, work for it.",
-  "Sweat is just fat crying.",
-  "You don't have to be extreme, just consistent.",
-  "The only way to define your limits is by going beyond them.",
-  "Your health is an investment, not an expense.",
+    description: "Burned over 5000 calories in cardio exercises"
+  }
 ];
